Refactor some e2e.go methods into main.go
